Move message listeners (FirebaseMessagingService implementation and didReceiveRemoteMessage delegate method) out of plugin into example app, as users need to implement these themselves - Remove unused platform code related to push notifications - Minor UI changes to reduce space usage
